Renamed the authentioncation state provider for the broker firm back office users.
